+++ ios/App/CapApp-SPM/Package.swift
@@ -15,6 +15,7 @@ let package = Package(
         .package(name: "CapacitorFilesystem", path: "../../../node_modules/@capacitor/filesystem"),
         .package(name: "CapacitorPreferences", path: "../../../node_modules/@capacitor/preferences"),
         .package(name: "CapacitorShare", path: "../../../node_modules/@capacitor/share"),
+        .package(name: "FoldExternalDisplay", path: "../../../native-plugins/fold-external-display"),
         .package(name: "FoldNativeCamera", path: "../../../native-plugins/fold-native-camera")
     ],
     targets: [
@@ -26,6 +27,7 @@ let package = Package(
                 .product(name: "CapacitorFilesystem", package: "CapacitorFilesystem"),
                 .product(name: "CapacitorPreferences", package: "CapacitorPreferences"),
                 .product(name: "CapacitorShare", package: "CapacitorShare"),
+                .product(name: "FoldExternalDisplay", package: "FoldExternalDisplay"),
                 .product(name: "FoldNativeCamera", package: "FoldNativeCamera")
             ]
         )